refactor, split exporter into multiple files ref #3
